Better method for determining if the canvas is blank
Just use a boolean flag to determine if the canvas has been edited rather than actually looking at the contents of the canvas. This works for every case except when they purposefully erase everything manually, but at that point we can't even try to anticipate what they're trying to do.

Resolves the remaining part of https://github.com/melindamorang/blowyourfaceoff/issues/9 to my satisfaction