Better way of determining when the game is over
Something unreliable was happening sometimes when trying to determine whether the game was over and whether we should go to the next round or to the endgame.  add-entry.php was taking care of this check and returning different text strings in its response to the javascript, and sometimes this was weird.  Scrap that and instead just store a variable in the Javascript with the number of rounds the game should have.  When the current round indicates it's the last round, move on to the endgame.

Add some plumbing to fetch and store the number of rounds there are based on the count of distinct players in the table.