Make it so you can draw single dots
I finally figured out what Jacob was talking about. In order to complete a line, including a zero-length line, you need to add the canvasContext.stroke() call. The problem before was that it was outside the if statement, so changing tools would cause random dots to get drawn on the canvas.  The cure for all our problems is to put this inside the if statement so it only completes the line when the mouse position is still within the canvas.