Show message and confirmation button if round times out with no input
If the timer runs out and the user has entered nothing, display a message on the page asking if the player is still there. To proceed, they have to click a button proving they are still paying attention to the tab. This prevents problems with multiple simultaneously open tabs submitting things at different times and overwriting each other. Note that it will also pause the game forever if the player has legitimately left, but if they did that, the game wouldn't be very fun anyway.

I initially tried using a Javascript confirm() alert, but I didn't like it because if you switch tabs, it goes away, and it also doesn't show in a non-active tab at all.  This homemade html alert seems more practical for the current purposes.