Fix errors on Android
I finally made the effort to get my phone set up where I can access my PC's localhost and where my PC can debug what's happening on my phone. Woohoo!

There was something funny happening in this piece of code where mouseEvent.touches[0] was actually undefined, and stuff after that was just failing, but it was working anyway. Weird. I don't know if any platform requires this little piece of code (maybe ios), but it should be safe to only update the mouseEvent if it's not undefined. With this change, undo now works on my Android.