Make the gameplay input zone not selectable
I think this will prevent weird drawing issues that occur when the canvas is inadvertently selected.

I tried to make only the canvas not selectable, but that didn't seem to work and I still had problems. This change here works better, although if the user selects the text above the canvas, you can still get in a weird selecty drag-drop state. But this seems to be a net improvement.